Updated server to be able to handle special server codes: when a message is sent with Server as its recipient (this nickname is reserved and cannot be set as their nickname by clients), that message will automatically be flagged as an operation and not a message. certain codes, followed by a zero, will trigger certain behaviours.
Currently this behaviour is limited to changing nicknames but it can certainly be expanded upon.

Next I'm looking to add the functionality for the server to set the local m_id in the client, as this is used as the sender when handling server codes.

I plan soon to make Message take two ints, being the senders Id and recepients ID, and handle who that message is addressed to and sent from internally within the server